More i64 -> i128

+++ esp-config/src/generate/validator.rs
@@ -16,7 +16,7 @@ pub enum Validator {
     /// Only allow positive integers, i.e. any values greater than to 0.
     PositiveInteger,
     /// Ensure that an integer value falls within the specified range.
-    IntegerInRange(Range<i64>),
+    IntegerInRange(Range<i128>),
     /// String-Enumeration. Only allows one of the given Strings.
     #[serde()]
     Enumeration(Vec<String>),
@@ -142,7 +142,7 @@ pub(crate) fn positive_integer(value: &Value) -> Result<(), Error> {
     Ok(())
 }
 
-pub(crate) fn integer_in_range(range: &Range<i64>, value: &Value) -> Result<(), Error> {
+pub(crate) fn integer_in_range(range: &Range<i128>, value: &Value) -> Result<(), Error> {
     if !value.is_integer() || !range.contains(&value.as_integer()) {
         Err(Error::validation(format!(
             "Value '{value}' does not fall within range '{range:?}'"
